fix: split route discovery loaders
