+++ src/routing/discovery/discover-route-modules.ts
@@ -1,0 +1,64 @@
+import type { RouteModule } from '@/Config';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+
+const supportedRouteModuleExtensions = new Set(['.js', '.mjs', '.cjs', '.ts', '.mts', '.cts']);
+
+export function discoverRouteModules(routesDir?: string): RouteModule[] {
+  if (routesDir === undefined) {
+    return [];
+  }
+
+  const rootDirectory = resolveModulePath(routesDir);
+
+  if (!fs.existsSync(rootDirectory)) {
+    throw new Error(`Routes directory does not exist: ${rootDirectory}`);
+  }
+
+  return readRouteModulePaths(rootDirectory).sort();
+}
+
+function readRouteModulePaths(currentDirectory: string): string[] {
+  const routeModules: string[] = [];
+  const entries = fs.readdirSync(currentDirectory, { withFileTypes: true });
+
+  for (const entry of entries) {
+    const absolutePath = path.join(currentDirectory, entry.name);
+
+    if (entry.isDirectory()) {
+      routeModules.push(...readRouteModulePaths(absolutePath));
+      continue;
+    }
+
+    if (isSupportedRouteModule(absolutePath)) {
+      routeModules.push(absolutePath);
+    }
+  }
+
+  return routeModules;
+}
+
+function isSupportedRouteModule(filePath: string): boolean {
+  if (filePath.endsWith('.d.ts')) {
+    return false;
+  }
+
+  if (hasIgnoredTestSuffix(filePath)) {
+    return false;
+  }
+
+  return supportedRouteModuleExtensions.has(path.extname(filePath));
+}
+
+function hasIgnoredTestSuffix(filePath: string): boolean {
+  return (
+    filePath.includes('.test.') ||
+    filePath.includes('.test-') ||
+    filePath.includes('.spec.') ||
+    filePath.includes('.spec-')
+  );
+}
+
+function resolveModulePath(modulePath: string, baseDirectory: string = process.cwd()): string {
+  return path.isAbsolute(modulePath) ? modulePath : path.resolve(baseDirectory, modulePath);
+}
